Refactoring Fragments.
