+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Users\Samim\.android\avd\Pixel_6_API_33.avd" />
+            <value value="C:\Users\Samim\.android\avd\Resizable_Experimental_API_33.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2023-04-27T13:01:38.862120900Z" />
+    <timeTargetWasSelectedWithDropDown value="2023-06-07T09:44:05.026511900Z" />
   </component>
 </project>
