+++ app/src/main/java/com/example/testproject/app/presentation/dashboard/DashboardFragment.kt
@@ -1,60 +1,115 @@
 package com.example.testproject.app.presentation.dashboard
 
+import android.content.Context
+import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.Fragment
+import android.util.Log
 import android.view.LayoutInflater
+import android.view.Menu
+import android.view.MenuInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import androidx.navigation.fragment.findNavController
 import com.example.testproject.R
+import com.example.testproject.app.presentation.app.App
+import com.example.testproject.app.presentation.factory.ViewModelFactory
+import com.example.testproject.databinding.FragmentDashboardBinding
+import com.google.android.material.snackbar.Snackbar
+import javax.inject.Inject
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [DashboardFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class DashboardFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
+    private var _binding: FragmentDashboardBinding? = null
+    private val binding: FragmentDashboardBinding
+        get() = _binding ?: throw RuntimeException("FragmentDashboardBinding == null")
+
+    @Inject
+    lateinit var viewModelFactory: ViewModelFactory
+    private val viewModel by lazy {
+        ViewModelProvider(this, viewModelFactory)[DashboardViewModel::class.java]
+    }
+
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        (context.applicationContext as App).component.inject(this@DashboardFragment)
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_dashboard, container, false)
+    ): View {
+        _binding = FragmentDashboardBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
-    companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment DashboardFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            DashboardFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        viewModel.firebaseUser.observe(viewLifecycleOwner) {
+            Log.d("Account user", "User observe activity: $it")
+            if (it == null) {
+                findNavController().navigate(R.id.action_dashboardFragment_to_loginFragment)
             }
+        }
+
+        appBarMenu()
+
+        binding.addScheduleButton.setOnClickListener {
+             launchNotificationFragment()
+        }
+        binding.cardClickToStart.setOnClickListener {
+            launchLevelFragment()
+        }
     }
+
+    private fun launchLevelFragment() {
+        findNavController().navigate(R.id.action_dashboardFragment_to_levelFragment)
+    }
+
+    private fun launchNotificationFragment() {
+        findNavController().navigate(R.id.action_dashboardFragment_to_notificationFragment)
+    }
+
+    private fun launchSettingsFragment() {
+        findNavController().navigate(R.id.action_dashboardFragment_to_settingsFragment)
+    }
+
+    private fun appBarMenu() {
+
+        binding.appBarLayout.setOnClickListener { view ->
+            Snackbar.make(view, "Replace with you own action", Snackbar.LENGTH_SHORT)
+                .setAction("Action", null).show()
+        }
+
+        binding.topAppBar.setOnMenuItemClickListener { menuItem ->
+            when (menuItem.itemId) {
+                R.id.alerts -> {
+                    Toast.makeText(requireContext(), "Alerts", Toast.LENGTH_SHORT).show()
+                    true
+                }
+                R.id.settings -> {
+                    launchSettingsFragment()
+                    true
+                }
+
+                else -> false
+            }
+        }
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+        super.onCreateOptionsMenu(menu, inflater)
+        inflater.inflate(R.menu.menu_dashboard, menu)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+
 }
