+++ app/src/main/java/com/example/testproject/app/presentation/main/MainActivity.kt
@@ -4,8 +4,6 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.example.testproject.R
 import com.example.testproject.app.presentation.app.App
-
-
 class MainActivity : AppCompatActivity() {
 
     private val component by lazy {
@@ -15,22 +13,6 @@ class MainActivity : AppCompatActivity() {
         component.inject(this)
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
-        /*
-        val notificationManager = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-
-        if (notificationManager.areNotificationsEnabled()) {
-            // Notifications are allowed, launch login activity
-            Handler().postDelayed(Runnable {
-                startActivity(LoginActivity.newIntent(this@MainActivity))
-                finish()
-            }, 5000)
-        } else {
-            // Notifications are not allowed, permission must be requested
-            val intent = Intent(Settings.ACTION_APP_NOTIFICATION_SETTINGS)
-            intent.putExtra(Settings.EXTRA_APP_PACKAGE, packageName)
-            startActivity(intent)
-        }*/
 
     }
