+++ app/src/main/java/com/example/testproject/app/presentation/registration/one/RegistrationOneFragment.kt
@@ -1,60 +1,123 @@
 package com.example.testproject.app.presentation.registration.one
 
+import android.content.Context
 import android.os.Bundle
-import androidx.fragment.app.Fragment
+import android.text.Editable
+import android.text.TextWatcher
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
 import com.example.testproject.R
+import com.example.testproject.app.presentation.app.App
+import com.example.testproject.app.utils.DataMask
+import com.example.testproject.databinding.FragmentRegistrationOneBinding
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [RegistrationOneFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class RegistrationOneFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
+    private var _binding: FragmentRegistrationOneBinding? = null
+    private val binding: FragmentRegistrationOneBinding
+        get() = _binding ?: throw RuntimeException("FragmentRegistrationOneBinding == null")
+
+    private val dataMask = DataMask()
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        (context.applicationContext as App).component.inject(this@RegistrationOneFragment)
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_registration_one, container, false)
+    ): View {
+        _binding = FragmentRegistrationOneBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+
+        binding.editDateOfBirth.addTextChangedListener(dataMask)
+        binding.buttonNextRegistration.isEnabled = false
+
+        enabledButton()
+
+        binding.buttonNextRegistration.setOnClickListener {
+            val name = binding.editTextNameRegistration.text?.trim().toString()
+            val lastName = binding.editTextLastNameRegistration.text?.trim().toString()
+            val date = binding.editDateOfBirth.text?.trim().toString()
+            Log.d("RegistrationActivity", "One activity extra: $name, $lastName, $date ")
+            launchTwoFragment(name, lastName, date)
+        }
+        binding.imageButtonArrowBack.setOnClickListener {
+            launchLoginFragment()
+        }
+    }
+
+    private fun launchLoginFragment() {
+        findNavController().navigate(R.id.action_registrationOneFragment_to_loginFragment)
+    }
+
+    private fun launchTwoFragment(name: String, lastName: String, date: String) {
+        val bundle = Bundle()
+        bundle.putString(PUT_NAME_KEY, name)
+        bundle.putString(PUT_LAST_NAME_KEY, lastName)
+        bundle.putString(PUT_DATE_OF_BIRTH_KEY, date)
+        findNavController().navigate(R.id.action_registrationOneFragment_to_registrationTwoFragment, bundle)
+    }
+
+
+    private fun enabledButton() {
+        binding.editTextNameRegistration.addTextChangedListener(object : TextWatcher {
+            override fun afterTextChanged(p0: Editable?) {
+                binding.buttonNextRegistration.isEnabled =
+                    ((binding.editTextNameRegistration.text?.length ?: 0) > 0
+                            && (binding.editTextLastNameRegistration.text?.length ?: 0) > 0
+                            && (binding.editDateOfBirth.text?.length ?: 0) == 10)
+
+            }
+
+            override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {}
+            override fun onTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {}
+        })
+
+        binding.editTextLastNameRegistration.addTextChangedListener(object : TextWatcher {
+            override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {}
+            override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {}
+            override fun afterTextChanged(s: Editable?) {
+                binding.buttonNextRegistration.isEnabled =
+                    ((binding.editTextLastNameRegistration.text?.length ?: 0) > 0
+                            && (binding.editTextNameRegistration.text?.length ?: 0) > 0
+                            && (binding.editDateOfBirth.text?.length ?: 0) == 10)
+            }
+
+        })
+
+        binding.editDateOfBirth.addTextChangedListener(object : TextWatcher {
+            override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {}
+            override fun onTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {}
+            override fun afterTextChanged(p0: Editable?) {
+                binding.buttonNextRegistration.isEnabled =
+                    ((binding.editDateOfBirth.text?.length ?: 0) == 10
+                            && (binding.editTextNameRegistration.text?.length ?: 0) > 0
+                            && (binding.editTextLastNameRegistration.text?.length ?: 0) > 0)
+            }
+        })
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
     }
 
     companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment RegistrationOneFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            RegistrationOneFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
-            }
+        private const val PUT_NAME_KEY = "name"
+        private const val PUT_LAST_NAME_KEY = "lastName"
+        private const val PUT_DATE_OF_BIRTH_KEY = "dateOfBirth"
     }
+
+
 }
