+++ app/src/main/java/com/example/testproject/app/presentation/registration/two/RegistrationTwoFragment.kt
@@ -1,60 +1,116 @@
 package com.example.testproject.app.presentation.registration.two
 
+import android.content.Context
 import android.os.Bundle
-import androidx.fragment.app.Fragment
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
 import com.example.testproject.R
+import com.example.testproject.app.presentation.app.App
+import com.example.testproject.databinding.FragmentRegistrationTwoBinding
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
 
-/**
- * A simple [Fragment] subclass.
- * Use the [RegistrationTwoFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class RegistrationTwoFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
+
+    private var _binding: FragmentRegistrationTwoBinding? = null
+    private val binding: FragmentRegistrationTwoBinding
+        get() = _binding ?: throw RuntimeException("FragmentRegistrationTwoBinding == null")
+
+    private var name: String? = null
+    private var lastName: String? = null
+    private var date: String? = null
+    private var gender = false
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        (context.applicationContext as App).component.inject(this@RegistrationTwoFragment)
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+            name = it.getString(GET_NAME_KEY)
+            lastName = it.getString(GET_LAST_NAME_KEY)
+            date = it.getString(GET_DATE_OF_BIRTH_KEY)
         }
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_registration_two, container, false)
+    ): View {
+        _binding = FragmentRegistrationTwoBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.buttonNextRegistration.isEnabled = false
+
+        if (name == null || lastName == null || date == null) {
+            launchRegistrationOneFragment()
+            return
+        } else {
+            Log.d("RegistrationActivity", "Two activity get: $name, $lastName, $date")
+            GenderObserver()
+        }
+    }
+
+    private fun GenderObserver() {
+        binding.cardViewMale.setOnClickListener {
+            binding.cardViewMale.isChecked = true
+            binding.cardViewFmale.isChecked = false
+            binding.buttonNextRegistration.isEnabled = true
+            gender = true
+        }
+
+        binding.cardViewFmale.setOnClickListener {
+            binding.cardViewFmale.isChecked = true
+            binding.cardViewMale.isChecked = false
+            binding.buttonNextRegistration.isEnabled = true
+            gender = false
+        }
+
+        binding.imageButtonArrowBack.setOnClickListener {
+            launchRegistrationOneFragment()
+        }
+
+        binding.buttonNextRegistration.setOnClickListener {
+            launchThreeFragment(name, lastName, date, gender)
+        }
+    }
+
+    private fun launchThreeFragment(name: String?, lastName: String?, date: String?, gender: Boolean) {
+        val bundle = Bundle()
+        bundle.putString(GET_NAME_KEY, name)
+        bundle.putString(GET_LAST_NAME_KEY, lastName)
+        bundle.putString(GET_DATE_OF_BIRTH_KEY, date)
+        bundle.putBoolean(PUT_GENDER_KEY, gender)
+        findNavController().navigate(
+            R.id.action_registrationTwoFragment_to_registrationThreeFragment,
+            bundle
+        )
+    }
+
+    private fun launchRegistrationOneFragment() {
+        findNavController().navigate(R.id.action_registrationTwoFragment_to_registrationOneFragment)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
     }
 
     companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment RegistrationTwoFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            RegistrationTwoFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
-            }
+        private const val GET_NAME_KEY = "name"
+        private const val GET_LAST_NAME_KEY = "lastName"
+        private const val GET_DATE_OF_BIRTH_KEY = "dateOfBirth"
+        private const val PUT_GENDER_KEY = "gender"
     }
+
+
 }
