+++ app/src/main/java/com/example/testproject/app/presentation/workout/list/ExerciseListFragment.kt
@@ -1,60 +1,150 @@
 package com.example.testproject.app.presentation.workout.list
 
+import android.content.Context
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.lifecycleScope
+import androidx.navigation.fragment.findNavController
+import com.bumptech.glide.Glide
 import com.example.testproject.R
+import com.example.testproject.app.common.Resource
+import com.example.testproject.app.domain.model.beginner.Exercise
+import com.example.testproject.app.presentation.app.App
+import com.example.testproject.app.presentation.factory.ViewModelFactory
+import com.example.testproject.app.presentation.workout.list.adapters.ExerciseAdapter
+import com.example.testproject.databinding.FragmentExerciseListBinding
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
+import javax.inject.Inject
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
 
-/**
- * A simple [Fragment] subclass.
- * Use the [ExerciseListFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class ExerciseListFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
+
+    private var _binding: FragmentExerciseListBinding? = null
+    private val binding: FragmentExerciseListBinding
+        get() = _binding ?: throw RuntimeException("FragmentExerciseListBinding == null")
+
+    @Inject
+    lateinit var viewModelFactory: ViewModelFactory
+    private val viewModel by lazy {
+        ViewModelProvider(this, viewModelFactory)[ExercisesViewModel::class.java]
+    }
+
+    private lateinit var exerciseAdapter: ExerciseAdapter
+
+    private var id: Int? = null
+    private var title: String? = null
+    private var picture: String? = null
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        (context.applicationContext as App).component.inject(this@ExerciseListFragment)
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+            id = it.getInt(GET_ID_KEY)
+            title = it.getString(GET_TITLE_KEY)
+            picture = it.getString(GET_PICTURE_KEY)
         }
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_exercise_list, container, false)
+    ): View {
+        _binding = FragmentExerciseListBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        if (id == null || title == null || picture == null) {
+            launchLevelFragment()
+            return
+        } else{
+            viewModel.loadExerciseList(id!!)
+            viewModel.exerciseInfoList.onEach {
+                when (it) {
+                    is Resource.Loading -> {
+                        binding.nestedScrollView.visibility = View.GONE
+                        binding.progressBar.visibility = View.VISIBLE
+                    }
+
+                    is Resource.Success -> {
+                        initAdapter(it)
+                        binding.textViewCountWorkout.text = it.data.size.toString()
+                        binding.textViewWorkout.text = title
+                        Glide.with(this)
+                            .load(picture)
+                            .centerCrop()
+                            .into(binding.imageViewWorkout)
+                        if (id!! < 6) {
+                            binding.ratingBarWorkout.rating = 1.0f
+                        } else if (id in 6..10) {
+                            binding.ratingBarWorkout.rating = 2.0f
+                        } else {
+                            binding.ratingBarWorkout.rating = 3.0f
+                        }
+                        binding.nestedScrollView.visibility = View.VISIBLE
+                        binding.progressBar.visibility = View.GONE
+                        exerciseAdapter.onExerciseClickListener = {
+                            launchDetailFragment(it.id, id!!)
+                        }
+                    }
+
+                    is Resource.Error -> {
+                        binding.nestedScrollView.visibility = View.GONE
+                        binding.progressBar.visibility = View.GONE
+                        Toast.makeText(requireContext(), it.message, Toast.LENGTH_SHORT).show()
+                    }
+                }
+            }.launchIn(lifecycleScope)
+        }
+
+        binding.imageViewArrowBack.setOnClickListener {
+            launchLevelFragment()
+        }
+    }
+
+    private fun launchLevelFragment() {
+        findNavController().navigate(R.id.action_exerciseListFragment_to_levelFragment)
+    }
+
+    private fun launchDetailFragment(idExercise: Int, idExerciseList: Int) {
+        val bundle = Bundle()
+        bundle.putInt(PUT_ID_EXERCISE_KEY, idExercise)
+        bundle.putInt(PUT_ID_EXERCISE_LIST_KEY, idExerciseList)
+        findNavController().navigate(R.id.action_exerciseListFragment_to_detailFragment, bundle)
+    }
+
+
+    private fun initAdapter(it: Resource.Success<List<Exercise>>) {
+        exerciseAdapter = ExerciseAdapter()
+        binding.recyclerViewExercise.adapter = exerciseAdapter
+        exerciseAdapter.submitList(it.data)
+
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
     }
 
     companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment ExerciseListFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            ExerciseListFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
-            }
+        private const val GET_ID_KEY = "id"
+        private const val GET_TITLE_KEY = "title"
+        private const val GET_PICTURE_KEY = "picture"
+        private const val PUT_ID_EXERCISE_KEY = "id_exercise"
+        private const val PUT_ID_EXERCISE_LIST_KEY = "id_exercise_list"
     }
+
+
 }
