--- app/src/main/res/layout/activity_detail.xml
@@ -1,156 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".app.presentation.workout.detail.DetailActivity">
-
-    <androidx.core.widget.NestedScrollView
-        android:id="@+id/nestedScrollView"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toTopOf="@+id/buttonClose"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="0dp">
-
-            <ProgressBar
-                android:id="@+id/progressBar"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:layout_constraintBottom_toBottomOf="@+id/cardAreaImage"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <com.google.android.material.card.MaterialCardView
-                android:id="@+id/cardAreaVideo"
-                android:layout_width="match_parent"
-                android:layout_height="250dp"
-                android:layout_margin="8dp"
-                android:backgroundTint="@color/white"
-                app:cardCornerRadius="10dp"
-                app:cardElevation="4dp"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:strokeWidth="0dp">
-
-                <com.pierfrancescosoffritti.androidyoutubeplayer.core.player.views.YouTubePlayerView
-                    android:id="@+id/youTubePlayerView"
-                    android:layout_width="match_parent"
-                    android:layout_height="250dp"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
-
-            </com.google.android.material.card.MaterialCardView>
-
-            <TextView
-                android:id="@+id/textViewTitle"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="8dp"
-                android:fontFamily="@font/ubuntu_medium"
-                android:gravity="center_vertical"
-                android:textColor="#757575"
-                android:textSize="24sp"
-                app:layout_constraintEnd_toEndOf="@+id/cardAreaVideo"
-                app:layout_constraintStart_toStartOf="@+id/cardAreaVideo"
-                app:layout_constraintTop_toBottomOf="@+id/cardAreaVideo"
-                tools:text="Jumping" />
-
-            <TextView
-                android:id="@+id/textViewSubtitle"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="8dp"
-                android:fontFamily="@font/ubuntu_medium"
-                android:gravity="center_vertical"
-                android:textColor="#94D647"
-                android:textSize="24sp"
-                app:layout_constraintEnd_toEndOf="@+id/textViewTitle"
-                app:layout_constraintStart_toStartOf="@+id/textViewTitle"
-                app:layout_constraintTop_toBottomOf="@+id/textViewTitle"
-                tools:text="16" />
-
-            <TextView
-                android:id="@+id/textViewDescription"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="8dp"
-                android:fontFamily="@font/ubuntu_regular"
-                android:textSize="16sp"
-                app:layout_constraintEnd_toEndOf="@+id/cardAreaVideo"
-                app:layout_constraintStart_toStartOf="@+id/cardAreaVideo"
-                app:layout_constraintTop_toBottomOf="@+id/textViewSubtitle"
-                tools:text="Description" />
-
-            <TextView
-                android:id="@+id/textViewAreaTitle"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="8dp"
-                android:fontFamily="@font/ubuntu_medium"
-                android:gravity="center_vertical"
-                android:text="@string/area_of_focus"
-                android:textColor="@android:color/holo_blue_light"
-                android:textSize="24sp"
-                app:layout_constraintEnd_toEndOf="@+id/textViewDescription"
-                app:layout_constraintStart_toStartOf="@+id/textViewDescription"
-                app:layout_constraintTop_toBottomOf="@+id/textViewDescription" />
-
-            <TextView
-                android:id="@+id/textViewAreaDescription"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="8dp"
-                android:fontFamily="@font/ubuntu_regular"
-                android:textSize="20sp"
-                app:layout_constraintEnd_toEndOf="@+id/textViewDescription"
-                app:layout_constraintStart_toStartOf="@+id/textViewAreaTitle"
-                app:layout_constraintTop_toBottomOf="@+id/textViewAreaTitle"
-                tools:text="Area description" />
-
-            <com.google.android.material.card.MaterialCardView
-                android:id="@+id/cardAreaImage"
-                android:layout_width="0dp"
-                android:layout_height="350dp"
-                android:layout_marginTop="8dp"
-                android:layout_marginBottom="8dp"
-                android:backgroundTint="@color/white"
-                app:cardCornerRadius="10dp"
-                app:cardElevation="4dp"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="@+id/textViewAreaDescription"
-                app:layout_constraintStart_toStartOf="@+id/textViewAreaDescription"
-                app:layout_constraintTop_toBottomOf="@+id/textViewAreaDescription"
-                app:strokeWidth="0dp">
-
-                <ImageView
-                    android:id="@+id/imageViewArea"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:scaleType="centerCrop" />
-
-            </com.google.android.material.card.MaterialCardView>
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-    </androidx.core.widget.NestedScrollView>
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/buttonClose"
-        android:layout_width="0dp"
-        android:layout_height="48dp"
-        android:text="@string/close"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="@+id/nestedScrollView"
-        app:layout_constraintStart_toStartOf="@+id/nestedScrollView" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
