--- app/src/main/res/layout/activity_reset.xml
@@ -1,137 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/green"
-    tools:context=".app.presentation.reset.ResetActivity">
-
-    <ImageView
-        android:id="@+id/imageButtonArrowBack"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="64dp"
-        android:src="@drawable/ic_arrow_back"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageView
-        android:id="@+id/textViewOneToFour"
-        android:layout_width="65dp"
-        android:layout_height="66dp"
-        android:layout_marginLeft="80dp"
-        android:background="@drawable/some_id"
-        android:elevation="12dp"
-        app:layout_constraintBottom_toTopOf="@+id/cardViewReg4"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/textViewTitleApp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:fontFamily="@font/ubuntu_medium"
-        android:text="@string/app_name"
-        android:textAppearance="@style/text_view_title_app"
-        android:textStyle="bold"
-        app:layout_constraintStart_toEndOf="@+id/textViewOneToFour"
-        app:layout_constraintTop_toTopOf="@+id/textViewOneToFour" />
-
-    <!-- Always pay -->
-    <TextView
-        android:id="@+id/textViewSubtitleApp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:fontFamily="@font/ubuntu_medium"
-        android:gravity="top"
-        android:text="@string/always_pay"
-        android:textAppearance="@style/text_view_subtitle_app"
-        app:layout_constraintStart_toStartOf="@+id/textViewTitleApp"
-        app:layout_constraintTop_toBottomOf="@+id/textViewTitleApp" />
-
-    <com.google.android.material.card.MaterialCardView
-        android:id="@+id/cardViewReg4"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:backgroundTint="@color/green"
-        app:cardCornerRadius="0dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:strokeWidth="0dp">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@drawable/rectangle_9">
-
-            <TextView
-                android:id="@+id/textViewReset"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="32dp"
-                android:text="@string/reset_your_"
-                android:textAppearance="@style/header"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <TextView
-                android:id="@+id/textViewPleaseEnter"
-                android:layout_width="310dp"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="25dp"
-                android:layout_marginTop="17dp"
-                android:layout_marginEnd="25dp"
-                android:gravity="center"
-                android:text="@string/please_ente"
-                android:textAppearance="@style/please_enter"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/textViewReset" />
-
-            <com.google.android.material.textfield.TextInputLayout
-                android:id="@+id/tilEmailReset"
-                style="@style/text_input_layout"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="25dp"
-                android:layout_marginTop="35dp"
-                android:layout_marginEnd="25dp"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintHorizontal_bias="0.5"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/textViewPleaseEnter">
-
-                <com.google.android.material.textfield.TextInputEditText
-                    android:id="@+id/editTextEmailLogin"
-                    style="@style/input_edit_text"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:hint="@string/email"
-                    android:inputType="textEmailAddress" />
-            </com.google.android.material.textfield.TextInputLayout>
-
-            <com.google.android.material.button.MaterialButton
-                android:id="@+id/buttonSend"
-                style="@style/OutlinedButton.Green"
-                android:layout_width="0dp"
-                android:layout_height="48dp"
-                android:layout_marginTop="24dp"
-                android:layout_marginBottom="64dp"
-                android:text="@string/send"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="@+id/tilEmailReset"
-                app:layout_constraintHorizontal_bias="0.5"
-                app:layout_constraintStart_toStartOf="@+id/tilEmailReset"
-                app:layout_constraintTop_toBottomOf="@+id/tilEmailReset" />
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-    </com.google.android.material.card.MaterialCardView>
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
