+++ app/src/main/res/navigation/main_navigation.xml
@@ -9,7 +9,7 @@
         android:id="@+id/splashFragment"
         android:name="com.example.testproject.app.presentation.splash.SplashFragment"
         android:label="fragment_splash"
-        tools:layout="@layout/fragment_splash" >
+        tools:layout="@layout/fragment_splash">
         <action
             android:id="@+id/action_splashFragment_to_loginFragment"
             app:destination="@id/loginFragment" />
@@ -18,7 +18,7 @@
         android:id="@+id/dashboardFragment"
         android:name="com.example.testproject.app.presentation.dashboard.DashboardFragment"
         android:label="fragment_dashboard"
-        tools:layout="@layout/fragment_dashboard" >
+        tools:layout="@layout/fragment_dashboard">
         <action
             android:id="@+id/action_dashboardFragment_to_notificationFragment"
             app:destination="@id/notificationFragment" />
@@ -28,12 +28,15 @@
         <action
             android:id="@+id/action_dashboardFragment_to_settingsFragment"
             app:destination="@id/settingsFragment" />
+        <action
+            android:id="@+id/action_dashboardFragment_to_loginFragment"
+            app:destination="@id/loginFragment" />
     </fragment>
     <fragment
         android:id="@+id/detailFragment"
         android:name="com.example.testproject.app.presentation.workout.detail.DetailFragment"
         android:label="fragment_detail"
-        tools:layout="@layout/fragment_detail" >
+        tools:layout="@layout/fragment_detail">
         <action
             android:id="@+id/action_detailFragment_to_exerciseListFragment"
             app:destination="@id/exerciseListFragment" />
@@ -42,7 +45,7 @@
         android:id="@+id/exerciseListFragment"
         android:name="com.example.testproject.app.presentation.workout.list.ExerciseListFragment"
         android:label="fragment_exercise_list"
-        tools:layout="@layout/fragment_exercise_list" >
+        tools:layout="@layout/fragment_exercise_list">
         <action
             android:id="@+id/action_exerciseListFragment_to_detailFragment"
             app:destination="@id/detailFragment" />
@@ -54,7 +57,7 @@
         android:id="@+id/levelFragment"
         android:name="com.example.testproject.app.presentation.workout.lvl.LevelFragment"
         android:label="fragment_level"
-        tools:layout="@layout/fragment_level" >
+        tools:layout="@layout/fragment_level">
         <action
             android:id="@+id/action_levelFragment_to_exerciseListFragment"
             app:destination="@id/exerciseListFragment" />
@@ -66,7 +69,7 @@
         android:id="@+id/loginFragment"
         android:name="com.example.testproject.app.presentation.login.LoginFragment"
         android:label="fragment_login"
-        tools:layout="@layout/fragment_login" >
+        tools:layout="@layout/fragment_login">
         <action
             android:id="@+id/action_loginFragment_to_registrationOneFragment"
             app:destination="@id/registrationOneFragment" />
@@ -81,51 +84,81 @@
         android:id="@+id/notificationFragment"
         android:name="com.example.testproject.app.presentation.notification.NotificationFragment"
         android:label="fragment_notification"
-        tools:layout="@layout/fragment_notification" />
+        tools:layout="@layout/fragment_notification">
+        <action
+            android:id="@+id/action_notificationFragment_to_dashboardFragment"
+            app:destination="@id/dashboardFragment" />
+        <action
+            android:id="@+id/action_notificationFragment_to_dashboardFragment2"
+            app:destination="@id/dashboardFragment" />
+    </fragment>
     <fragment
         android:id="@+id/registrationFourFragment"
         android:name="com.example.testproject.app.presentation.registration.four.RegistrationFourFragment"
         android:label="fragment_registration_four"
-        tools:layout="@layout/fragment_registration_four" >
+        tools:layout="@layout/fragment_registration_four">
         <action
             android:id="@+id/action_registrationFourFragment_to_loginFragment"
             app:destination="@id/loginFragment" />
+        <action
+            android:id="@+id/action_registrationFourFragment_to_registrationThreeFragment"
+            app:destination="@id/registrationThreeFragment" />
     </fragment>
     <fragment
         android:id="@+id/registrationOneFragment"
         android:name="com.example.testproject.app.presentation.registration.one.RegistrationOneFragment"
         android:label="fragment_registration_one"
-        tools:layout="@layout/fragment_registration_one" >
+        tools:layout="@layout/fragment_registration_one">
         <action
             android:id="@+id/action_registrationOneFragment_to_registrationTwoFragment"
             app:destination="@id/registrationTwoFragment" />
+        <action
+            android:id="@+id/action_registrationOneFragment_to_loginFragment"
+            app:destination="@id/loginFragment" />
     </fragment>
     <fragment
         android:id="@+id/registrationThreeFragment"
         android:name="com.example.testproject.app.presentation.registration.three.RegistrationThreeFragment"
         android:label="fragment_registration_three"
-        tools:layout="@layout/fragment_registration_three" >
+        tools:layout="@layout/fragment_registration_three">
         <action
             android:id="@+id/action_registrationThreeFragment_to_registrationFourFragment"
             app:destination="@id/registrationFourFragment" />
+        <action
+            android:id="@+id/action_registrationThreeFragment_to_registrationTwoFragment"
+            app:destination="@id/registrationTwoFragment" />
     </fragment>
     <fragment
         android:id="@+id/registrationTwoFragment"
         android:name="com.example.testproject.app.presentation.registration.two.RegistrationTwoFragment"
         android:label="fragment_registration_two"
-        tools:layout="@layout/fragment_registration_two" >
+        tools:layout="@layout/fragment_registration_two">
         <action
             android:id="@+id/action_registrationTwoFragment_to_registrationThreeFragment"
             app:destination="@id/registrationThreeFragment" />
+        <action
+            android:id="@+id/action_registrationTwoFragment_to_registrationOneFragment"
+            app:destination="@id/registrationOneFragment" />
     </fragment>
     <fragment
         android:id="@+id/resetFragment"
         android:name="com.example.testproject.app.presentation.reset.ResetFragment"
         android:label="fragment_reset"
-        tools:layout="@layout/fragment_reset" />
+        tools:layout="@layout/fragment_reset">
+        <action
+            android:id="@+id/action_resetFragment_to_loginFragment"
+            app:destination="@id/loginFragment" />
+    </fragment>
     <fragment
         android:id="@+id/settingsFragment"
         android:name="com.example.testproject.app.presentation.settings.SettingsFragment"
         android:label="fragment_settings"
-        tools:layout="@layout/fragment_settings" />
+        tools:layout="@layout/fragment_settings">
+        <action
+            android:id="@+id/action_settingsFragment_to_loginFragment"
+            app:destination="@id/loginFragment" />
+        <action
+            android:id="@+id/action_settingsFragment_to_dashboardFragment"
+            app:destination="@id/dashboardFragment" />
+    </fragment>
 </navigation>
